update naming for api
