--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/SouthernBlueRingedOctopusConfiguration.java
@@ -1,13 +1,0 @@
-package com.adthena.southernblueringedoctopus;
-
-import io.dropwizard.Configuration;
-import io.dropwizard.db.DataSourceFactory;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-public class SouthernBlueRingedOctopusConfiguration extends Configuration {
-
-  private DataSourceFactory database;
-}
