--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/api/Category.java
@@ -1,13 +1,0 @@
-package com.adthena.southernblueringedoctopus.api;
-
-import com.adthena.southernblueringedoctopus.db.entities.CategoryEntity;
-import java.util.List;
-
-public interface Category {
-
-  CategoryEntity getCategoryById(Integer catid);
-
-  CategoryEntity upsert(CategoryEntity category);
-
-  List<CategoryEntity> list();
-}
