--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/api/Date.java
@@ -1,10 +1,0 @@
-package com.adthena.southernblueringedoctopus.api;
-
-import com.adthena.southernblueringedoctopus.db.entities.DateEntity;
-
-public interface Date {
-
-  DateEntity getDateById(Integer dateid);
-
-  DateEntity upsert(DateEntity date);
-}
