--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/api/Event.java
@@ -1,10 +1,0 @@
-package com.adthena.southernblueringedoctopus.api;
-
-import com.adthena.southernblueringedoctopus.db.entities.EventEntity;
-
-public interface Event {
-
-  EventEntity getEventById(Integer eventid);
-
-  EventEntity upsert(EventEntity event);
-}
