--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/api/Venue.java
@@ -1,11 +1,0 @@
-package com.adthena.southernblueringedoctopus.api;
-
-import com.adthena.southernblueringedoctopus.db.entities.VenueEntity;
-import java.util.List;
-
-public interface Venue {
-
-  VenueResource getVenueById(Integer venueid);
-
-  List<VenueEntity> list();
-}
