--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/db/dao/VenueDao.java
@@ -1,17 +1,0 @@
-package com.adthena.southernblueringedoctopus.db.dao;
-
-import com.adthena.southernblueringedoctopus.db.entities.VenueEntity;
-import java.util.List;
-import org.jdbi.v3.sqlobject.config.RegisterConstructorMapper;
-import org.jdbi.v3.sqlobject.statement.SqlQuery;
-
-public interface VenueDao {
-
-  @RegisterConstructorMapper(VenueEntity.class)
-  @SqlQuery("SELECT * FROM venue WHERE venueid=:venueid")
-  VenueEntity getVenueById(Integer venueid);
-
-  @SqlQuery("SELECT * FROM venue ORDER BY venueid ASC")
-  @RegisterConstructorMapper(VenueEntity.class)
-  List<VenueEntity> list();
-}
