--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/db/entities/CategoryEntity.java
@@ -1,14 +1,0 @@
-package com.adthena.southernblueringedoctopus.db.entities;
-
-import lombok.Value;
-import lombok.With;
-
-@Value
-public class CategoryEntity {
-
-  @With
-  Integer catid;
-  String catgroup;
-  String catname;
-  String catdesc;
-}
