--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/db/entities/DateEntity.java
@@ -1,21 +1,0 @@
-package com.adthena.southernblueringedoctopus.db.entities;
-
-import java.time.LocalDate;
-import lombok.ToString;
-import lombok.Value;
-import lombok.With;
-
-@Value
-@ToString
-public class DateEntity {
-
-  @With
-  Integer dateid;
-  LocalDate caldate;
-  String day;
-  Integer week;
-  String month;
-  String qtr;
-  Integer year;
-  Boolean holiday;
-}
