--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/db/entities/EventEntity.java
@@ -1,17 +1,0 @@
-package com.adthena.southernblueringedoctopus.db.entities;
-
-import java.time.LocalDateTime;
-import lombok.Value;
-import lombok.With;
-
-@Value
-public class EventEntity {
-
-  @With
-  Integer eventid;
-  Integer venueid;
-  Integer catid;
-  Integer dateid;
-  String eventname;
-  LocalDateTime starttime;
-}
