--- lead_fullstack_developer_test/southern-blue-ringed-octopus/src/main/java/com/adthena/southernblueringedoctopus/db/entities/VenueEntity.java
@@ -1,13 +1,0 @@
-package com.adthena.southernblueringedoctopus.db.entities;
-
-import lombok.Value;
-
-@Value
-public class VenueEntity {
-
-  Integer venueid;
-  String venuename;
-  String venuecity;
-  String venuestate;
-  Integer venueseats;
-}
