+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/TestApiConfiguration.java
@@ -1,0 +1,13 @@
+package com.adthena.testapi;
+
+import io.dropwizard.Configuration;
+import io.dropwizard.db.DataSourceFactory;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+public class TestApiConfiguration extends Configuration {
+
+  private DataSourceFactory database;
+}
