+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/api/Category.java
@@ -1,0 +1,13 @@
+package com.adthena.testapi.api;
+
+import com.adthena.testapi.db.entities.CategoryEntity;
+import java.util.List;
+
+public interface Category {
+
+  CategoryEntity getCategoryById(Integer catid);
+
+  CategoryEntity upsert(CategoryEntity category);
+
+  List<CategoryEntity> list();
+}
