+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/api/Date.java
@@ -1,0 +1,10 @@
+package com.adthena.testapi.api;
+
+import com.adthena.testapi.db.entities.DateEntity;
+
+public interface Date {
+
+  DateEntity getDateById(Integer dateid);
+
+  DateEntity upsert(DateEntity date);
+}
