+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/api/Event.java
@@ -1,0 +1,10 @@
+package com.adthena.testapi.api;
+
+import com.adthena.testapi.db.entities.EventEntity;
+
+public interface Event {
+
+  EventEntity getEventById(Integer eventid);
+
+  EventEntity upsert(EventEntity event);
+}
