+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/api/Venue.java
@@ -1,0 +1,11 @@
+package com.adthena.testapi.api;
+
+import com.adthena.testapi.db.entities.VenueEntity;
+import java.util.List;
+
+public interface Venue {
+
+  VenueResource getVenueById(Integer venueid);
+
+  List<VenueEntity> list();
+}
