+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/db/dao/VenueDao.java
@@ -1,0 +1,17 @@
+package com.adthena.testapi.db.dao;
+
+import com.adthena.testapi.db.entities.VenueEntity;
+import java.util.List;
+import org.jdbi.v3.sqlobject.config.RegisterConstructorMapper;
+import org.jdbi.v3.sqlobject.statement.SqlQuery;
+
+public interface VenueDao {
+
+  @RegisterConstructorMapper(VenueEntity.class)
+  @SqlQuery("SELECT * FROM venue WHERE venueid=:venueid")
+  VenueEntity getVenueById(Integer venueid);
+
+  @SqlQuery("SELECT * FROM venue ORDER BY venueid ASC")
+  @RegisterConstructorMapper(VenueEntity.class)
+  List<VenueEntity> list();
+}
