+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/db/entities/CategoryEntity.java
@@ -1,0 +1,14 @@
+package com.adthena.testapi.db.entities;
+
+import lombok.Value;
+import lombok.With;
+
+@Value
+public class CategoryEntity {
+
+  @With
+  Integer catid;
+  String catgroup;
+  String catname;
+  String catdesc;
+}
