+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/db/entities/DateEntity.java
@@ -1,0 +1,21 @@
+package com.adthena.testapi.db.entities;
+
+import java.time.LocalDate;
+import lombok.ToString;
+import lombok.Value;
+import lombok.With;
+
+@Value
+@ToString
+public class DateEntity {
+
+  @With
+  Integer dateid;
+  LocalDate caldate;
+  String day;
+  Integer week;
+  String month;
+  String qtr;
+  Integer year;
+  Boolean holiday;
+}
