+++ lead_fullstack_developer_test/test-api/src/main/java/com/adthena/testapi/db/entities/VenueEntity.java
@@ -1,0 +1,13 @@
+package com.adthena.testapi.db.entities;
+
+import lombok.Value;
+
+@Value
+public class VenueEntity {
+
+  Integer venueid;
+  String venuename;
+  String venuecity;
+  String venuestate;
+  Integer venueseats;
+}
