+++ lead_fullstack_developer_test/test-api/src/main/resources/db/migration/V1__populate_db.sql
